+++ app/src/main/java/com/sostrovsky/travelup/domain/settings/SettingsDomain.kt
@@ -5,4 +5,4 @@ package com.sostrovsky.travelup.domain.settings
  * Date: 12.09.20
  * Email: sergey.ostrovsky.it.dev@gmail.com
  */
-data class SettingsDomain(val languageId: Long, val currencyId: Long, val countryId: Long)
+data class SettingsDomain(val languageId: Int, val currencyId: Int, val countryId: Int)
